+++ src/config/platform.php
@@ -1,0 +1,71 @@
+<?php
+
+/**
+ * This file is part of Laravel Core by Graham Campbell.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at http://bit.ly/UWsjkb.
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+return [
+
+    /*
+    |--------------------------------------------------------------------------
+    | Site Name
+    |--------------------------------------------------------------------------
+    |
+    | This defines the site name.
+    |
+    | Default to 'Laravel Application'.
+    |
+    */
+
+    'name' => 'Laravel Application',
+
+    /*
+    |--------------------------------------------------------------------------
+    | Site Brand
+    |--------------------------------------------------------------------------
+    |
+    | This defines the site brand.
+    |
+    | Default to 'Laravel Web Services'.
+    |
+    */
+
+    'brand' => 'Laravel Web Services',
+
+    /*
+    |--------------------------------------------------------------------------
+    | Site Author
+    |--------------------------------------------------------------------------
+    |
+    | This defines the site author.
+    |
+    | Default to 'Author'.
+    |
+    */
+
+    'author' => 'Author',
+
+    /*
+    |--------------------------------------------------------------------------
+    | Site Description
+    |--------------------------------------------------------------------------
+    |
+    | This defines the site author.
+    |
+    | Default to 'Description.'.
+    |
+    */
+
+    'description' => 'Description.',
+
+];